optimize job manager and executor
